LPS-98618 Inline ConfigurationIndexingExtender into SearchMVCRenderCommand, no logic change
