+++ modules/apps/configuration-admin/configuration-admin-web/src/main/java/com/liferay/configuration/admin/web/internal/portlet/action/SearchMVCRenderCommand.java
@@ -22,33 +22,49 @@ import com.liferay.configuration.admin.web.internal.display.ConfigurationModelCo
 import com.liferay.configuration.admin.web.internal.display.ConfigurationScreenConfigurationEntry;
 import com.liferay.configuration.admin.web.internal.display.context.ConfigurationScopeDisplayContext;
 import com.liferay.configuration.admin.web.internal.model.ConfigurationModel;
+import com.liferay.configuration.admin.web.internal.search.ConfigurationModelIndexer;
 import com.liferay.configuration.admin.web.internal.search.FieldNames;
 import com.liferay.configuration.admin.web.internal.util.ConfigurationEntryIterator;
 import com.liferay.configuration.admin.web.internal.util.ConfigurationEntryRetriever;
+import com.liferay.configuration.admin.web.internal.util.ConfigurationModelIterator;
 import com.liferay.configuration.admin.web.internal.util.ConfigurationModelRetriever;
 import com.liferay.configuration.admin.web.internal.util.ResourceBundleLoaderProvider;
 import com.liferay.portal.configuration.metatype.annotations.ExtendedObjectClassDefinition;
+import com.liferay.portal.kernel.cluster.ClusterMasterExecutor;
+import com.liferay.portal.kernel.log.Log;
+import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.CompanyConstants;
 import com.liferay.portal.kernel.portlet.bridges.mvc.MVCRenderCommand;
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.Hits;
+import com.liferay.portal.kernel.search.IndexWriterHelper;
 import com.liferay.portal.kernel.search.Indexer;
 import com.liferay.portal.kernel.search.IndexerRegistry;
 import com.liferay.portal.kernel.search.QueryConfig;
 import com.liferay.portal.kernel.search.SearchContext;
+import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.concurrent.ConcurrentHashMap;
 
 import javax.portlet.PortletException;
 import javax.portlet.RenderRequest;
 import javax.portlet.RenderResponse;
 
+import org.osgi.framework.Bundle;
+import org.osgi.framework.BundleContext;
+import org.osgi.framework.BundleEvent;
+import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Deactivate;
 import org.osgi.service.component.annotations.Reference;
+import org.osgi.util.tracker.BundleTracker;
+import org.osgi.util.tracker.BundleTrackerCustomizer;
 
 /**
  * @author Michael C. Han
@@ -177,16 +193,162 @@ public class SearchMVCRenderCommand implements MVCRenderCommand {
 		return "/search_results.jsp";
 	}
 
+	@Activate
+	protected void activate(BundleContext bundleContext) {
+		if (_clusterMasterExecutor.isMaster()) {
+			Bundle[] bundles = bundleContext.getBundles();
+
+			List<ConfigurationModel> configurationModelsList =
+				new ArrayList<>();
+
+			for (Bundle bundle : bundles) {
+				if (bundle.getState() != Bundle.ACTIVE) {
+					continue;
+				}
+
+				Map<String, ConfigurationModel> configurationModels =
+					_configurationModelRetriever.getConfigurationModels(
+						bundle, ExtendedObjectClassDefinition.Scope.SYSTEM,
+						null);
+
+				configurationModelsList.addAll(configurationModels.values());
+
+				_configurationModelsMap.put(
+					bundle.getSymbolicName(), configurationModels.values());
+			}
+
+			_configurationModelIndexer.reindex(configurationModelsList);
+
+			commit(_configurationModelIndexer);
+		}
+
+		_bundleTracker = new BundleTracker<>(
+			bundleContext, Bundle.ACTIVE,
+			new ConfigurationModelsBundleTrackerCustomizer());
+
+		_bundleTracker.open();
+	}
+
+	protected void commit(Indexer<ConfigurationModel> indexer) {
+		try {
+			_indexWriterHelper.commit(indexer.getSearchEngineId());
+		}
+		catch (SearchException se) {
+			if (_log.isWarnEnabled()) {
+				_log.warn("Unable to commit", se);
+			}
+		}
+	}
+
+	@Deactivate
+	protected void deactivate() {
+		_bundleTracker.close();
+
+		_bundleTracker = null;
+	}
+
+	private static final Log _log = LogFactoryUtil.getLog(
+		SearchMVCRenderCommand.class);
+
+	private BundleTracker<ConfigurationModelIterator> _bundleTracker;
+
+	@Reference
+	private ClusterMasterExecutor _clusterMasterExecutor;
+
 	@Reference
 	private ConfigurationEntryRetriever _configurationEntryRetriever;
 
 	@Reference
+	private ConfigurationModelIndexer _configurationModelIndexer;
+
+	@Reference
 	private ConfigurationModelRetriever _configurationModelRetriever;
+
+	private final Map<String, Collection<ConfigurationModel>>
+		_configurationModelsMap = new ConcurrentHashMap<>();
 
 	@Reference
 	private IndexerRegistry _indexerRegistry;
 
 	@Reference
+	private IndexWriterHelper _indexWriterHelper;
+
+	@Reference
 	private ResourceBundleLoaderProvider _resourceBundleLoaderProvider;
+
+	private class ConfigurationModelsBundleTrackerCustomizer
+		implements BundleTrackerCustomizer<ConfigurationModelIterator> {
+
+		@Override
+		public ConfigurationModelIterator addingBundle(
+			Bundle bundle, BundleEvent bundleEvent) {
+
+			if (!_clusterMasterExecutor.isMaster()) {
+				return null;
+			}
+
+			Collection<ConfigurationModel> configurationModels =
+				_configurationModelsMap.remove(bundle.getSymbolicName());
+
+			if (configurationModels != null) {
+				if (configurationModels.isEmpty()) {
+					return null;
+				}
+
+				return new ConfigurationModelIterator(configurationModels);
+			}
+
+			Map<String, ConfigurationModel> configurationModelsMap =
+				_configurationModelRetriever.getConfigurationModels(
+					bundle, ExtendedObjectClassDefinition.Scope.SYSTEM, null);
+
+			if (configurationModelsMap.isEmpty()) {
+				return null;
+			}
+
+			_configurationModelIndexer.reindex(configurationModelsMap.values());
+
+			commit(_configurationModelIndexer);
+
+			return new ConfigurationModelIterator(
+				configurationModelsMap.values());
+		}
+
+		@Override
+		public void modifiedBundle(
+			Bundle bundle, BundleEvent bundleEvent,
+			ConfigurationModelIterator configurationModelIterator) {
+
+			removedBundle(bundle, bundleEvent, configurationModelIterator);
+
+			addingBundle(bundle, bundleEvent);
+		}
+
+		@Override
+		public void removedBundle(
+			Bundle bundle, BundleEvent bundleEvent,
+			ConfigurationModelIterator configurationModelIterator) {
+
+			if (!_clusterMasterExecutor.isMaster()) {
+				return;
+			}
+
+			for (ConfigurationModel configurationModel :
+					configurationModelIterator.getResults()) {
+
+				try {
+					_configurationModelIndexer.delete(configurationModel);
+				}
+				catch (SearchException se) {
+					if (_log.isWarnEnabled()) {
+						_log.warn("Unable to reindex models", se);
+					}
+				}
+			}
+
+			commit(_configurationModelIndexer);
+		}
+
+	}
 
 }
